chore: move test cases to swc
