+++ crates/swc/tests/fixture/interop/none/input/.swcrc
@@ -1,0 +1,12 @@
+{
+    "jsc": {
+        "parser": {
+            "syntax": "ecmascript"
+        },
+        "target": "es2020"
+    },
+    "module": {
+        "type": "commonjs",
+        "importInterop": "none"
+    }
+}
