+++ crates/swc/tests/fixture/interop/none/output/index.js
@@ -1,0 +1,23 @@
+"use strict";
+function _export(target, all) {
+    for(var name in all)Object.defineProperty(target, name, {
+        get: all[name],
+        enumerable: true
+    });
+}
+_export(exports, {
+    default: ()=>_default,
+    x: ()=>x,
+    y: ()=>y
+});
+const _foo = require("foo");
+const _bar = require("bar");
+const x = {
+    foo: _foo.default,
+    bar: _bar.default
+};
+const y = _bar.baz;
+var _default = {
+    x,
+    y
+};
