+++ crates/swc/tests/fixture/interop/swc/input/index.js
@@ -1,0 +1,9 @@
+import foo from "foo";
+import bar, { baz } from "bar";
+
+export const x = { foo, bar };
+export const y = baz;
+export default {
+    x,
+    y,
+};
